feat: mock show command
