+++ cli/tests/catalog_responses/hello_show.json
@@ -1,0 +1,18 @@
+[
+  {
+    "results": [
+      {
+        "input": "nixpkgs",
+        "system": "aarch64-darwin",
+        "relPath": [
+          "hello"
+        ],
+        "pname": "hello",
+        "version": "2.12.1",
+        "description": "A program that produces a familiar, friendly greeting",
+        "license": "GPL-3.0-or-later"
+      }
+    ],
+    "count": 1
+  }
+]
